refactor(orders): share the execution skeleton across order types
